+++ es6.js
@@ -31,7 +31,7 @@ module.exports = {
     ecmaVersion: 6
   },
   rules: {
-    'class-methods-use-this': 'warn',
+    'class-methods-use-this': 'off',
     'no-empty-pattern': 'error',
     'prefer-promise-reject-errors': 'error',
     'template-tag-spacing': [